slight tweaks to README and run_analysis.r
Clarified that the script CAN be run as an executable script but that it
was INTENDED to be run line by line with some human review check points.

Also added a print statement at the end of run_analysis.r to notify the
user that the file completed running (in case it is run as an executable
script).